requirements.txt
aiosqlite==0.20.0
aiogram==3.12.0
python-decouple==3.8
sqlalchemy==2.0.35
requests==2.32.3